old-tests/submission/Infraware/Forms/contents/Forms/select_getterindex.html

<!DOCTYPE HTML>
<html>
 <head>
  <title>Forms</title>
  <script src="../../resources/testharness.js"></script>
  <script src="../../resources/testharnessreport.js"></script>
 </head>
 <body>
	<p>
		<h3>select_getterindex</h3>
	</p>
	
	<hr>

	<div id="log"></div>

	<form method="post"
		  enctype="application/x-www-form-urlencoded"
		  action=""
		  id="input_form">
		<p>
		 <label for="unittype">Select unit type:</label>
		 <select id="unittype" name="unittype">
		  <option value="1"> Miner </option>
		  <option value="2"> Puffer </option>
		  <option value="3" selected> Snipey </option>
		  <option value="4"> Max </option>
		  <option value="5"> Firebot </option>
		 </select>
		</p>	
	</form>

	<script>
		
		var select = document.getElementById("unittype");
		var item = select[2];

		if (typeof(item) == "object") {
			test(function() {
				assert_equals(item.value, "3", "calling of getterindex method is failed.");
			});
		} else {
			test(function() {
				assert_unreached("calling of getterindex method is failed.");
			});
		}

	</script>

 </body>
</html>
